+++ ruoyi-fastapi-backend/module_ai/entity/do/ai_chat_do.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import CHAR, BigInteger, Column, DateTime, Float, Integer, Text
+from sqlalchemy import CHAR, Integer, Column, DateTime, Float, Integer, Text
 
 from config.database import Base
 
@@ -13,8 +13,8 @@ class AiChatConfig(Base):
     __tablename__ = 'ai_chat_config'
     __table_args__ = {'comment': 'AI对话配置表'}
 
-    chat_config_id = Column(BigInteger, primary_key=True, nullable=False, autoincrement=True, comment='配置主键')
-    user_id = Column(BigInteger, nullable=False, unique=True, comment='用户ID')
+    chat_config_id = Column(Integer, primary_key=True, nullable=False, autoincrement=True, comment='配置主键')
+    user_id = Column(Integer, nullable=False, unique=True, comment='用户ID')
     temperature = Column(Float, nullable=True, comment='默认温度')
     add_history_to_context = Column(CHAR(1), server_default='0', comment='是否添加历史记录(0是, 1否)')
     num_history_runs = Column(Integer, nullable=True, comment='历史记录条数')
